Commits Tab Migration and Dev Email Lowercase
